Refactoring depency injection in ServiceController and AnimeService
